feat(#367): mock api.presentation
